Add commonFundingMockData.ts Begin refactoring existing tests
Ticket: AT-6739
